\name{pwt1001}
\docType{data}
\alias{pwt1001}
\title{Penn World Table (Version 10.01)}
\description{
  Relative levels of income, output, input, and productivity for 183 countries between 1950 and 2019
(base year: 2017).
}
\usage{pwt1001}
\format{A data frame with 12,810 observations of 52 variables.}
\details{The Penn World Table 10.01 (<https://www.rug.nl/ggdc/productivity/pwt/>) provides information on relative levels of income, output, input, and productivity for 183 countries between 1950 and 2019.
  }
\references{
  Morales-Oñate, V., and Morales-Oñate, B. (2024). \emph{Cluster Evolution Analytics}. \url{https://mpra.ub.uni-muenchen.de/120220/}
}
\keyword{datasets}