\name{actpas}
\docType{data}
\alias{actpas}
\title{Assets and liabilities operations}
\description{
  Ecuador's amount of Assets and Liabilities Operations of the National Financial System: https://contenido.bce.fin.ec/home1/economia/tasas/IndiceSFN.htm
}
\usage{actpas}
\format{A dataframe containing 358 observations and 25 columns.}
\references{
  Morales-Oñate, V., and Morales-Oñate, B. (2024). \emph{Cluster Evolution Analytics}. \doi{10.1016/j.softx.2024.101921}
}
\keyword{datasets}